Support NetworkManager
